% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plots.R
\name{add_legend}
\alias{add_legend}
\title{add_legend}
\usage{
add_legend(legend_location = "bottomright", ...)
}
\arguments{
\item{legend_location}{Network centrality score matrix}

\item{...}{Numeric, complex, or logical vectors.}
}
\value{
a legend to a plot
}
\description{
Add plot legend and locate it outside of a network plot
}
