% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{calculate_centrality_scores}
\alias{calculate_centrality_scores}
\title{calculate_centrality_scores}
\usage{
calculate_centrality_scores(net, centrality_type = "Degree")
}
\arguments{
\item{net}{protein-protein interaction network of igraph}

\item{centrality_type}{the maximum number of clusters}
}
\value{
A vector of network centrality scores
}
\description{
'calculate_centrality_scores' computes network centrality scores
}
