% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plots.R
\name{default_size}
\alias{default_size}
\title{default_size}
\usage{
default_size(max_cluster)
}
\arguments{
\item{max_cluster}{the maximum number of clusters}
}
\value{
A vector of plot width and height
}
\description{
Calculate image size based on the number of clusters
}
