% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_clusters}
\alias{example_clusters}
\title{Randomly generated example datasets for XINA users.
A dataset containing the XINA clustering results.
example_clusters}
\format{A list with the example XINA clustering result}
\description{
\itemize{
  \item aligned. XINA clustering results aligned by conditions
  \item data_column. Column names for data matrix
  \item out_dir. Not available in this example dataset
  \item nClusters. The number of user-desired clusters.
  It's 30 in the example.
  \item max_cluster. The number of clusters found in the dataset.
  It's 21 in the example.
  \item chosen_model. The chosen covariance model
  for the example dataset.  It's VEI in the example
  \item optimal_BIC. BIC at the optimized clustering.
  It's 29473.57 in the example
  \item condition. The experimental conditions in the dataset.
  \item color_for_condition. The default color
  for the conditions that will be used in XINA plot drawing.
  \item color_for_clusters. The default color
  for the clusters that will be used in XINA clustering plot.
  \item norm_method. The used normalization method
  to standardize the input data. It's "sum_normalization"
  in the example.
}
}
