% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{extract_data_column}
\alias{extract_data_column}
\title{extract_data_column}
\usage{
extract_data_column(col_head_of_clustering)
}
\arguments{
\item{col_head_of_clustering}{Column names of XINA clustering result}
}
\value{
A vector containing column names of  data matrix
}
\description{
Extract data column names from XINA clustering result
}
