% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{find_similar_clusters}
\alias{find_similar_clusters}
\title{find_similar_clusters}
\usage{
find_similar_clusters(clustering_result, threshold = 0.95)
}
\arguments{
\item{clustering_result}{A list containing XINA clustering results.
See \link[XINA]{xina_clustering}}

\item{threshold}{Pearson's r threshold to find similar ones}
}
\value{
Write a csv file containing similar clustering information
based on the given Pearson's R threshold
}
\description{
Compare clusters and find similar ones
}
