% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coregulstions.R
\name{generate_count_table}
\alias{generate_count_table}
\title{generate_count_table}
\usage{
generate_count_table(clustering_result, selected_conditions,
  comigration_size)
}
\arguments{
\item{clustering_result}{A list containing XINA clustering results. See \link[XINA]{xina_clustering}}

\item{selected_conditions}{A vector of condition names used in XINA clustering results.}

\item{comigration_size}{The number of proteins comigrated together in the selected conditions of XINA clustering results. Default is 0.}
}
\value{
A data frame containing comigrations.
}
\description{
Count the number of comigrated proteins using \link[plyr]{count}
}
