% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{generate_superset}
\alias{generate_superset}
\title{generate_superset}
\usage{
generate_superset(f_names, data_column, delim = ",",
  norm = "sum_normalization")
}
\arguments{
\item{f_names}{A vector of .csv file paths containing kinetics data}

\item{data_column}{A vector of column names containing data matrix}

\item{delim}{The delimiter of input file (default is ',')}

\item{norm}{The normalization method. It should be one of
c('sum_normalization', 'zscore').
Default is 'sum_normalization'.}
}
\value{
A data frame containing kinetics data obtained from files
in the f_names vector
}
\description{
Merge input kinetics files
}
