% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{get_colors}
\alias{get_colors}
\title{get_colors}
\usage{
get_colors(nClusters, set = "", colorset = NULL)
}
\arguments{
\item{nClusters}{The number of clusters}

\item{set}{Pre-defined color series set}

\item{colorset}{manually defined color codes}
}
\value{
A vector for color code of XINA graphics
}
\description{
Generate color series for XINA graphics
}
