% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plots.R
\name{get_layout}
\alias{get_layout}
\title{get_layout}
\usage{
get_layout(subnet_condition)
}
\arguments{
\item{subnet_condition}{A igraph sub-network}
}
\value{
igraph network layout
}
\description{
Get igraph layout by the number of nodes
}
