% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{get_mTOR_proteins}
\alias{get_mTOR_proteins}
\title{get_mTOR_proteins}
\usage{
get_mTOR_proteins(time_points, conditions)
}
\arguments{
\item{time_points}{A vector containing time points of the data matrix}

\item{conditions}{A vector containing condition information, for example normal, disease and drug treated disase.}
}
\value{
A vector containing mTOR pathway gene names
}
\description{
Get mTOR pathway genes
}
