% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{get_random_data}
\alias{get_random_data}
\title{get_random_data}
\usage{
get_random_data(time_points, conditions, num_total, percent.sign = 0.1,
  equal = TRUE)
}
\arguments{
\item{time_points}{A vector containing time points of the data matrix}

\item{conditions}{A vector containing condition information,
for example normal, disease and drug treated disase.}

\item{num_total}{The number of total proteins to be generated}

\item{percent.sign}{Percentage of differentially expressed proteins.
Ignored when equal=FALSE.}

\item{equal}{If equal is TRUE, all the conditions will have numbers
between 0 and 1.
If it is  FALSE, the first three conditions will have different ranges.
First condition will have numbers from 0.3 to 0.4.
Second condition will have numbers from 0.6 to 0.8.
Third condition will have numbers from 0.3 to 0.5.
Other conditions will have numbers from 0 to 1.}
}
\value{
A list containing ramdomly generated data matrix
}
\description{
Get randomized time-series data
}
