% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{get_stats}
\alias{get_stats}
\title{get_stats}
\usage{
get_stats(centrality_results, na.rm = FALSE)
}
\arguments{
\item{centrality_results}{Network centrality score data frame calculated by XINA network module}

\item{na.rm}{If it is FALSE, no exclusion of NA values.}
}
\value{
A data frame containing statistics of XINA network centrality scores
}
\description{
Calculate statistics of the given data for XINA network analysis
}
