% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{get_theme_blank}
\alias{get_theme_blank}
\title{get_theme_blank}
\usage{
get_theme_blank()
}
\value{
A ggplot theme
}
\description{
Predefined ggplot theme for removing ticks, titles and labels of X and Y axis
}
