% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{get_unknown_ppi_nodes}
\alias{get_unknown_ppi_nodes}
\title{get_unknown_ppi_nodes}
\usage{
get_unknown_ppi_nodes(xina_result, cl)
}
\arguments{
\item{xina_result}{A list containing XINA network analysis results. See \link[XINA]{xina_analysis}}

\item{cl}{the clustering number of XINA clustering results. See \link[XINA]{xina_clustering}}
}
\value{
A data frame containing proteins with no known interactions within the cluster based on the used protein-protein interaction database source
}
\description{
Get proteins with no known interactions within the cluster based on the used protein-protein interaction database source
}
\examples{
library(STRINGdb)

# load XINA example data
data(xina_example)

# Get STRING database for protein-protein intereaction information
string_db <- STRINGdb$new( version='10', species=9606, score_threshold=0, input_directory='' )
string_db

# XINA analysis with STRING DB
xina_result <- xina_analysis(example_clusters, string_db, flag_simplify=FALSE)

# Extract unknown PPI nodes in the cluster #1
get_unknown_ppi_nodes(xina_result, 1)

}
