% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gn}
\alias{gn}
\title{A character vector containing 19,396 human genes
This is for the randome data generation of XINA}
\format{A character vector containing 19,396 human genes}
\source{
\url{https://www.ncbi.nlm.nih.gov/gene}
}
\description{
\itemize{
  \item Characters of human genes
}
}
