% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gn_desc}
\alias{gn_desc}
\title{A character vector containing 19,396 human gene descriptions
This is for the randome data generation of XINA}
\format{A character vector containing 19,396 human gene descriptions}
\source{
\url{https://www.ncbi.nlm.nih.gov/gene}
}
\description{
\itemize{
  \item Human gene description corresponding to 'gn' vector
}
}
