% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hprd_ppi}
\alias{hprd_ppi}
\title{Protein-protein interaction resource downloaded from HPRD DB
A data frame containing HRPD protein-protein interaction data
hprd_ppi}
\format{A data frame containing HRPD protein-protein interaction data}
\source{
\url{http://www.hprd.org/}
}
\description{
\itemize{
  \item gene_symbol_1. Gene name interacting with gene name
  in 'gene_symbol_2'
  \item gene_symbol_2. Gene name interacting with gene name
  in 'gene_symbol_1'
  \item Experiment_type. Experimental or computational methods
  supporting the interaction
}
}
