% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{length2}
\alias{length2}
\title{length2}
\usage{
length2(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{If it is FALSE, no exclusion of NA values.}
}
\value{
A vector length
}
\description{
Customized function for vector length calculation
}
