% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{load_previous_results}
\alias{load_previous_results}
\title{load_previous_results}
\usage{
load_previous_results(clustering_dir = getwd(), data_column = NULL,
  fp_clusters = "xina_clusters.csv")
}
\arguments{
\item{clustering_dir}{The directory path of XINA clustering results}

\item{data_column}{A vector containing column names of  data matrix}

\item{fp_clusters}{File path of XINA clustering results}
}
\value{
Comma-separated file containing aligned XINA clustering results.
}
\description{
Get previous XINA clustering results to R space
}
\examples{

# Generate random multiplexed time-series data
random_data_info <- make_random_xina_data()

# Data files
data_files <- paste(random_data_info$conditions, ".csv", sep='')

# time points of the data matrix
data_column <- random_data_info$time_points

# mclust requires the fixed random seed to get reproduce the clustering results
set.seed(0)

# Run the model-based clustering
example_clusters <- xina_clustering(data_files, data_column=data_column, nClusters=30)

# Reload the clustering result
example_clusters_reloaded <- load_previous_results(".")

}
