% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{organize_clusters}
\alias{organize_clusters}
\title{organize_clusters}
\usage{
organize_clusters(clustering_dir = getwd(), super_ds, file_out = TRUE)
}
\arguments{
\item{clustering_dir}{The directory path of XINA clustering results}

\item{super_ds}{XINA clusters}

\item{file_out}{If it is TRUE, it writes the aligned clustering informaion
to "xina_clusters_aligned.csv" file.}
}
\value{
Comma-separated file containing aligned XINA clustering results.
}
\description{
Organize XINA clustering information by gene name
}
