% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{rank_centrality}
\alias{rank_centrality}
\title{rank_centrality}
\usage{
rank_centrality(centrality_score, type, num_breaks = 5)
}
\arguments{
\item{centrality_score}{Network centrality score matrix}

\item{type}{Network centrality score type, such as 'Eigenvector'}

\item{num_breaks}{The number of ranks}
}
\value{
A vector containing ranks
}
\description{
Give ranks based on network centrality scores
}
