% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{string_example}
\alias{string_example}
\title{Protein-protein interaction resource downloaded from STRING DB
for XINA's example dataset
A data frame containing protein-protein interactions}
\format{A data frame containing STRING protein-protein interaction data}
\source{
\url{https://string-db.org/}
}
\description{
\itemize{
  \item gene_symbol_1. Gene name interacting with gene name
  in 'gene_symbol_2'
  \item gene_symbol_2. Gene name interacting with gene name
  in 'gene_symbol_1'
  \item PPI_Source. Data original source
}
}
