% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{xina_result_example}
\alias{xina_result_example}
\title{Previously processed xina analysis using XINA's random example data
A list containing 'xina_analysis' results}
\format{A data frame containing STRING protein-protein interaction data}
\source{
XINA
}
\description{
\itemize{
  \item All_network. PPI network of all the input proteins
  \item Sub_network. A list containing PPI networks of each clusters
  \item Data. XINA clustering results. See \link[XINA]{xina_clustering}
  \item Nodes. A list of proteins in each cluster
  \item Conditions. A list of experimental condition of proteins
  in each cluster
  \item Titles.  A list of plot titles for XINA plotting
  \item out_dir. A directory path storing XINA network analysis results
  \item is_stringdb. False = different PPI DB and TRUE = STRING DB
}
}
