% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xina.R
\name{load_previous_results}
\alias{load_previous_results}
\title{load_previous_results}
\usage{
load_previous_results(clustering_dir = getwd(), data_column = NULL,
  fp_clusters = "xina_clusters.csv")
}
\arguments{
\item{clustering_dir}{The directory path of XINA clustering results}

\item{data_column}{A vector containing column names of  data matrix}

\item{fp_clusters}{File path of XINA clustering results}
}
\value{
Comma-separated file containing aligned XINA clustering results.
}
\description{
Get previous XINA clustering results to R space
}
\examples{
# Load XINA's example data
data(xina_example)
write.csv(example_clusters$aligned,"xina_clusters_aligned.csv")
write.csv(example_clusters$clusters,"xina_clusters.csv")

# Reload the clustering result
example_clusters_reloaded <- load_previous_results(".")

}
